+++ admin/class-sessions-admin.php
@@ -350,6 +350,7 @@ class Sessions_Admin {
 				Option::network_set( 'analytics', array_key_exists( 'pose_plugin_features_analytics', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_analytics' ) : false );
 				Option::network_set( 'metrics', array_key_exists( 'pose_plugin_features_metrics', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_metrics' ) : false );
 				Option::network_set( 'forceip', array_key_exists( 'pose_plugin_features_forceip', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_forceip' ) : false );
+				Option::network_set( 'killonreset', array_key_exists( 'pose_plugin_features_killonreset', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_killonreset' ) : false );
 				Option::network_set( 'followip', array_key_exists( 'pose_plugin_features_followip', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_followip' ) : false );
 				Option::network_set( 'history', array_key_exists( 'pose_plugin_features_history', $_POST ) ? (string) filter_input( INPUT_POST, 'pose_plugin_features_history', FILTER_SANITIZE_NUMBER_INT ) : Option::network_get( 'history' ) );
 				Option::network_set( 'rolemode', array_key_exists( 'pose_plugin_features_rolemode', $_POST ) ? (string) filter_input( INPUT_POST, 'pose_plugin_features_rolemode', FILTER_SANITIZE_NUMBER_INT ) : Option::network_get( 'rolemode' ) );
@@ -699,6 +700,22 @@ class Sessions_Admin {
 			]
 		);
 		register_setting( 'pose_plugin_features_section', 'pose_plugin_features_followip' );
+		add_settings_field(
+			'pose_plugin_features_killonreset',
+			esc_html__( 'On password reset', 'sessions' ),
+			[ $form, 'echo_field_checkbox' ],
+			'pose_plugin_features_section',
+			'pose_plugin_features_section',
+			[
+				'text'        => esc_html__( 'Delete current sessions', 'sessions' ),
+				'id'          => 'pose_plugin_features_killonreset',
+				'checked'     => Option::network_get( 'killonreset' ),
+				'description' => esc_html__( 'If checked, Sessions will delete all sessions of users resetting their passwords (recommended).', 'sessions' ),
+				'full_width'  => false,
+				'enabled'     => true,
+			]
+		);
+		register_setting( 'pose_plugin_features_section', 'pose_plugin_features_killonreset' );
 		add_settings_field(
 			'pose_plugin_features_analytics',
 			esc_html__( 'Analytics', 'sessions' ),