+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [2.6.0] - Not Yet Released
 
+### Added
+- A new option to delete all sessions of users resetting their passwords (thanks to [mrdexters1](https://profiles.wordpress.org/mrdexters1/) for the suggestion).
+
 ### Changed
 - Updated DecaLog SDK from version 2.0.0 to version 2.0.2.
 - Updated PerfOps One library from 2.2.1 to 2.2.2.